+++ pychron/core/regression/base_regressor.py
@@ -26,8 +26,8 @@ from traits.api import Array, List, Event, Property, Any, \
 from pychron.core.regression.tinv import tinv
 from pychron.core.stats.core import calculate_mswd, validate_mswd
 from pychron.core.stats.monte_carlo import RegressionEstimator
+from pychron.core.utils import alphas
 from pychron.pychron_constants import PLUSMINUS
-from pychron.utils import alphas
 
 logger = logging.getLogger('BaseRegressor')